improved and cleaned up + NVRAM and Saves idea
https://retropie.org.uk/forum/topic/26909/sega-model-3-segamodel3/129

Changes :

moved the emulator binary to /opt/retropie/emulators/supermodel-mechafatnick
reduced the run-lines, removing the path to the emulator binary, now ./supermodel
removed the -no-dsb in the underclock run-lines
it will preserve the changes made in Supermodel.ini when emulator is reinstalled
it will move the NVRAM and Saves directories from the source code to /opt/retropie/emulators/supermodel-mechafatnick
the initial save state uses now the %BASENAME%.st in /opt/retropie/emulators/supermodel-mechafatnick/Saves
An example of the initial save state is:

you make a savestate from, for example, scud.zip, it will become scudau.st0
then rename it to scud.st
on next game boot scud.st will be loaded as initial state
So basically if all NVRAM.nv and all Saves.st files come in the source-code.
No settings will have to be made and we can skip the initial boot sequence, for the roms with a slow startup.

+++ 00-scriptmodules-00/emulators/supermodel-mechafatnick.sh
@@ -10,7 +10,7 @@
 #
 
 rp_module_id="supermodel-mechafatnick"
-rp_module_desc="Sega supermodel 3 (Pi-friendly by Mechafatnick)"
+rp_module_desc="Sega supermodel 3 version by Mechafatnick and Pi-friendly"
 rp_module_help="\
 WARNING: Still under construction\n\n\
 Notes:\n\
@@ -49,11 +49,22 @@ function build_supermodel-mechafatnick() {
 }
 
 function install_supermodel-mechafatnick() {
-    md_ret_files=(        
-        'SuperModelPi-master/bin'
-        'SuperModelPi-master/Config'
-    )
+    #note: it should remove all files in $md_inst if there are any, so no need to remove links or files
     #note: command "mv" and "cp" won't work here, you get a "cannot stat error" in this function
+    #prevent an error if the NVRAM and Saves directory are not in the source!
+    if [[ -d "$md_build/SuperModelPi-master/NVRAM" ]] && [[ -d "$md_build/SuperModelPi-master/Saves" ]] ;then
+    md_ret_files=(        
+        'SuperModelPi-master/bin/supermodel'
+        'SuperModelPi-master/Config'
+        'SuperModelPi-master/NVRAM'
+        'SuperModelPi-master/Saves'
+    )
+    else
+    md_ret_files=(        
+        'SuperModelPi-master/bin/supermodel'
+        'SuperModelPi-master/Config'
+    ) 
+    fi
 }
 
 function configure_supermodel-mechafatnick() {
@@ -63,29 +74,31 @@ function configure_supermodel-mechafatnick() {
     mkRomDir "supermodel/model3emu/mechafatnick/NVRAM"
     mkRomDir "supermodel/model3emu/mechafatnick/Saves"    
 
-    addEmulator 0 "Supermodel-mechafatnick-normal" "supermodel" "XINIT:pushd /opt/retropie/emulators/supermodel-mechafatnick;/opt/retropie/emulators/supermodel-mechafatnick/bin/supermodel -wide-screen -legacy3d -sound-volume=50 -music-volume=60 -no-vsync -no-throttle -no-dsb -load-state=%ROM%.st0 %ROM%"
-    addEmulator 0 "Supermodel-mechafatnick-40-hz-PPC-Underclock" "supermodel" "XINIT:pushd /opt/retropie/emulators/supermodel-mechafatnick;/opt/retropie/emulators/supermodel-mechafatnick/bin/supermodel -wide-screen -ppc-frequency=40 -legacy3d -sound-volume=50 -music-volume=60 -no-vsync -no-throttle -no-dsb -load-state=%ROM%.st0 %ROM%"
-    addEmulator 0 "Supermodel-mechafatnick-45-hz-PPC-Underclock" "supermodel" "XINIT:pushd /opt/retropie/emulators/supermodel-mechafatnick;/opt/retropie/emulators/supermodel-mechafatnick/bin/supermodel -wide-screen -ppc-frequency=45 -legacy3d -sound-volume=50 -music-volume=60 -no-vsync -no-throttle -no-dsb -load-state=%ROM%.st0 %ROM%"
-    addEmulator 0 "Supermodel-mechafatnick-48-hz-PPC-Underclock" "supermodel" "XINIT:pushd /opt/retropie/emulators/supermodel-mechafatnick;/opt/retropie/emulators/supermodel-mechafatnick/bin/supermodel -wide-screen -ppc-frequency=48 -legacy3d -sound-volume=50 -music-volume=60 -no-vsync -no-throttle -no-dsb -load-state=%ROM%.st0 %ROM%"
+    addEmulator 0 "Supermodel-mechafatnick-normal" "supermodel" "XINIT:pushd /opt/retropie/emulators/supermodel-mechafatnick;./supermodel -wide-screen -legacy3d -sound-volume=50 -music-volume=60 -no-vsync -no-throttle -no-dsb -load-state=$romdir/supermodel/model3emu/mechafatnick/Saves/%BASENAME%.st %ROM%"
+    addEmulator 0 "Supermodel-mechafatnick-40-hz-PPC-Underclock" "supermodel" "XINIT:pushd /opt/retropie/emulators/supermodel-mechafatnick;./supermodel -wide-screen -ppc-frequency=40 -legacy3d -sound-volume=50 -music-volume=60 -no-vsync -no-throttle -load-state=$romdir/supermodel/model3emu/mechafatnick/Saves/%BASENAME%.st %ROM%"
+    addEmulator 0 "Supermodel-mechafatnick-45-hz-PPC-Underclock" "supermodel" "XINIT:pushd /opt/retropie/emulators/supermodel-mechafatnick;./supermodel -wide-screen -ppc-frequency=45 -legacy3d -sound-volume=50 -music-volume=60 -no-vsync -no-throttle -load-state=$romdir/supermodel/model3emu/mechafatnick/Saves/%BASENAME%.st %ROM%"
+    addEmulator 0 "Supermodel-mechafatnick-48-hz-PPC-Underclock" "supermodel" "XINIT:pushd /opt/retropie/emulators/supermodel-mechafatnick;./supermodel -wide-screen -ppc-frequency=48 -legacy3d -sound-volume=50 -music-volume=60 -no-vsync -no-throttle -load-state=$romdir/supermodel/model3emu/mechafatnick/Saves/%BASENAME%.st %ROM%"
 
     addSystem "supermodel" "Sega supermodel 3" ".zip"
 
-    #to prevent nested links within the directories, previously created links are removed before creating new ones !
-    rm "$md_inst/NVRAM" 2>&-
-    rm "$md_inst/Saves" 2>&-
-
+    #using "sudo ./retropie_packages.sh supermodel-mechafatnick install" for testing this function,
+    #you have to prevent nested links within the linked directories !
+    #if the link NVRAM doesn't exist, others will too, like: Saves,bin/Supermodel.ini and Supermodel.log
+    #then we can assume the installed directories are in place and we can move them and create symlinks
+    #else we do nothing otherwise we are going to push links into links!
+    if [[ ! -L "$md_inst/NVRAM" ]];then
+    mv "$md_inst/NVRAM" "$romdir/supermodel/model3emu/mechafatnick/NVRAM" 2>&-
+    mv "$md_inst/Saves" "$romdir/supermodel/model3emu/mechafatnick/Saves" 2>&-
+    #we don't want to overwrite an existing Supermodel.ini
+    [[ ! -f "$romdir/supermodel/model3emu/mechafatnick/Supermodel.ini" ]] && mv "$md_inst/Config/Supermodel.ini" "$romdir/supermodel/model3emu/mechafatnick/Supermodel.ini"
+    #If it can't be copied then remove te file, otherwise we can't make a symlink
+    [[ -f "$romdir/supermodel/model3emu/mechafatnick/Supermodel.ini" ]] &&rm "$md_inst/Config/Supermodel.ini"
+    #fix permissions
+    chown $user:$user -R "$romdir/supermodel/model3emu/mechafatnick"
+    #make symlinks
     ln -sv "$romdir/supermodel/model3emu/mechafatnick/NVRAM" "$md_inst/NVRAM"
     ln -sv "$romdir/supermodel/model3emu/mechafatnick/Saves" "$md_inst/Saves"
-    ln -sv "$romdir/supermodel/model3emu/mechafatnick/Supermodel.log" "$md_inst/Supermodel.log" 2>&-
-
-    #check if file exists and check if file is a not symlink
-    #if so, then : 
-    #- move the Supermodel.ini to the desired directory
-    #- fix the permissions of that file to the normal user
-    #- make a symlink to that file
-    if [[ -f "$md_inst/Config/Supermodel.ini" ]] && [[ ! -L "$md_inst/Config/Supermodel.ini" ]]; then
-    mv "$md_inst/Config/Supermodel.ini" "$romdir/supermodel/model3emu/mechafatnick/Supermodel.ini"
-    chown $user:$user "$romdir/supermodel/model3emu/mechafatnick/Supermodel.ini"
     ln -sv "$romdir/supermodel/model3emu/mechafatnick/Supermodel.ini" "$md_inst/Config/Supermodel.ini"
+    ln -sv "$romdir/supermodel/model3emu/mechafatnick/Supermodel.log" "$md_inst/Supermodel.log" 2>&-
     fi
 }
